Add comments for data reader and adapt the model according to parallel reader.
